Extract `PollingExecutorScheduler` abstraction
And implement `QueueExecutorScheduler` in terms of it, and make it the default
