+++ core/native/src/main/scala/cats/effect/unsafe/IORuntimeCompanionPlatform.scala
@@ -20,10 +20,9 @@ import scala.concurrent.ExecutionContext
 
 private[unsafe] abstract class IORuntimeCompanionPlatform { this: IORuntime.type =>
 
-  def defaultComputeExecutionContext: ExecutionContext =
-    scala.scalanative.runtime.ExecutionContext.global
+  def defaultComputeExecutionContext: ExecutionContext = QueueExecutorScheduler
 
-  def defaultScheduler: Scheduler = GlobalBusyWaitScheduler
+  def defaultScheduler: Scheduler = QueueExecutorScheduler
 
   private[this] var _global: IORuntime = null
